Cliente v1

+++ server/controllers/posts.controllers.js
@@ -15,10 +15,10 @@ export const createPost = async (req, res) => {
   try {
     let image;
 
-    if (req.files.image) {
+    if (req.files?.image && req.files.image !== null) {
       const { secure_url, public_id } = await uploadImage(req.files.image.tempFilePath);
       await fs.remove(req.files.image.tempFilePath);
-      image = {
+      req.body.image = {
         url: secure_url,
         public_id: public_id,
       };
